Display Autocomplete on Mobile
The autocomplete was previously hidden on mobile devices since there
were no styling rules applied for mobile out of the box. This is
somewhat of a limitation as the plugin doesn't provide any "example" to
follow if you're not using a theme that supports a mobile autocomplete.
On mobile devices, the autocomplete will now render in a slightly
different layout, with trending/related searches appearing on top of the
product results to better suit a portrait style view. Landscape view
will not be supported as you can't really see anything in the autocomplete
anyway.

SEARCHAUTO-9